// src/enums/rank.enum.ts
export enum RankTier {
  LVL_1 = 'LVL 1',
  LVL_2 = 'LVL 2',
  LVL_3 = 'LVL 3',
  LVL_4 = 'LVL 4',
  LVL_5 = 'LVL 5',
  MAX = 'MAX'
}

export enum RankTitle {
  NOVICE = 'GÀ TẬP SỰ',
  WARRIOR = 'CHIẾN BINH PHỔI SẠCH',
  KNIGHT = 'DŨNG SĨ HƠI THỞ XANH',
  MASTER = 'ĐẠI HIỆP KHÍ THUẦN',
  GRANDMASTER = 'BẬC THẦY KHÔNG KHÓI',
  LEGEND = 'HUYỀN THOẠI BẤT TỬ'
}